Move formats to file.
Took 9 minutes

Took 4 seconds
